+++ frontend/src/components/app/app.module.css
@@ -1,4 +1,4 @@
-.conteiner {
+.container {
   margin-top: 156px;
   margin-left: 394px;
 }
