+++ frontend/src/components/button/button.tsx
@@ -1,0 +1,1 @@
+// Сделать компонент Button со вскми вариациями кнопки
